refactor: subject 데이터 관리하는 클래스들 패키지 이동
