+++ src/main/java/kr/allcll/backend/admin/subject/TargetSubjectService.java
@@ -1,4 +1,4 @@
-package kr.allcll.backend.admin.seat;
+package kr.allcll.backend.admin.subject;
 
 import java.util.List;
 import java.util.Map;
@@ -8,7 +8,6 @@ import kr.allcll.backend.admin.seat.dto.PinSubjectUpdateRequest.PinSubject;
 import kr.allcll.crawler.common.exception.CrawlerAllcllException;
 import kr.allcll.crawler.subject.CrawlerSubject;
 import kr.allcll.crawler.subject.CrawlerSubjectRepository;
-import kr.allcll.backend.admin.subject.SubjectFilter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
